LayoutTests/imported/w3c/web-platform-tests/referrer-policy/generic/sandboxed-iframe-with-opaque-origin-expected.txt

Referrer Policy: A document with an opaque origin doesn't send referrers



PASS Sandboxed iframe with opaque origin doesn't send referrers to subresources
FAIL Sandboxed iframe with tuple origin sends referrers to subresources assert_equals: expected (string) "http://localhost:8800/referrer-policy/generic/sandboxed-iframe-with-opaque-origin.html" but got (undefined) undefined
FAIL Sandboxed iframe with opaque origin doesn't send referrers on navigation assert_equals: expected "" but got "http://localhost:8800/referrer-policy/generic/sandboxed-iframe-with-opaque-origin.html"
PASS Sandboxed iframe with tuple origin sends referrers on navigation
